Added JavaDoc to ShowConsolePageAction and moved it
