+++ org.eclipse.buildship.ui/src/main/java/org/eclipse/buildship/ui/view/execution/ExecutionPage.java
@@ -40,6 +40,7 @@ import org.eclipse.buildship.ui.PluginImages;
 import org.eclipse.buildship.ui.generic.CollapseTreeNodesAction;
 import org.eclipse.buildship.ui.generic.ExpandTreeNodesAction;
 import org.eclipse.buildship.ui.generic.GotoTestElementAction;
+import org.eclipse.buildship.ui.generic.ShowConsolePageAction;
 import org.eclipse.buildship.ui.view.BasePage;
 import org.eclipse.buildship.ui.view.MultiPageView;
 import org.eclipse.buildship.ui.view.PageSite;
